+++ apis/nucleus/src/components/listbox/assets/__tests__/measure-text.test.js
@@ -28,4 +28,10 @@ describe('measure-text', () => {
     expect(text).toEqual('');
     expect(text).toHaveLength(0);
   });
+
+  it('should handle a number instead of extracting it from layout', () => {
+    const text = getMeasureText(9);
+    expect(text).toEqual('MMMMMMMMM');
+    expect(text).toHaveLength(9);
+  });
 });