+++ apis/nucleus/src/components/listbox/components/__tests__/list-box-search.test.jsx
@@ -33,7 +33,7 @@ describe('<ListBoxSearch />', () => {
   beforeEach(() => {
     InstanceContextModule.default = InstanceContext;
 
-    store = { getStoreValue: jest.fn() };
+    store = { getStoreValue: jest.fn(), setStoreValue: jest.fn() };
     store.getStoreValue.mockImplementation((key) => {
       switch (key) {
         case 'listCount':