+++ apis/nucleus/src/components/listbox/components/grid-list-components/frequency-allowed.js
@@ -3,6 +3,6 @@ export const FREQUENCY_MIN_SHOW_WIDTH = 80;
 export default function getFrequencyAllowed({ itemWidth, layout, frequencyMode }) {
   const widthPermitsFreq = itemWidth > FREQUENCY_MIN_SHOW_WIDTH;
   const { frequencyEnabled = false } = layout?.qListObject || {};
-  const hasValidFreqOption = !['none', undefined].includes(frequencyMode);
+  const hasValidFreqOption = !['N', undefined].includes(frequencyMode);
   return !!(widthPermitsFreq && (hasValidFreqOption || frequencyEnabled));
 }