+++ apis/supernova/src/hooks.js
@@ -164,9 +164,9 @@ export function runSnaps(component, layout) {
   return Promise.resolve();
 }
 
-export function runMenu(component, menu, event) {
+export function runMenu(component, menu, event, menuBuilder) {
   try {
-    return Promise.all(component.__hooks.menus.map((h) => Promise.resolve(h.fn(menu, event)))).then(
+    return Promise.all(component.__hooks.menus.map((h) => Promise.resolve(h.fn(menu, event, menuBuilder)))).then(
       (menus) => menus[menus.length - 1]
     );
   } catch (e) {
@@ -1114,10 +1114,15 @@ export function useRenderState() {
 }
 
 /**
+ * @class Emitter
+ * @description The emitter instance. Implements https://nodejs.org/api/events.html#class-eventemitter.
+ */
+
+/**
  * Gets an event emitter instance for the visualization.
  * @entry
  * @experimental
- * @returns {Emitter} The emitter instance. Implements https://nodejs.org/api/events.html#class-eventemitter.
+ * @returns {Emitter}
  * @example
  * // In a Nebula visualization
  * import { useEmitter } from '@nebula.js/stardust';