+++ apis/supernova/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.3.0](https://github.com/qlik-oss/nebula.js/compare/v4.2.4...v4.3.0) (2023-08-03)
+
+**Note:** Version bump only for package @nebula.js/supernova
+
+## [4.2.4](https://github.com/qlik-oss/nebula.js/compare/v4.2.3...v4.2.4) (2023-07-04)
+
+### Features
+
+- add menuBuilder ([#1298](https://github.com/qlik-oss/nebula.js/issues/1298)) ([413db1b](https://github.com/qlik-oss/nebula.js/commit/413db1b66868c1a488dfa5d2b2462971ea7293ff))
+
+## [4.2.3](https://github.com/qlik-oss/nebula.js/compare/v4.2.2...v4.2.3) (2023-06-28)
+
+**Note:** Version bump only for package @nebula.js/supernova
+
+## [4.2.2](https://github.com/qlik-oss/nebula.js/compare/v4.2.1...v4.2.2) (2023-06-28)
+
+**Note:** Version bump only for package @nebula.js/supernova
+
+## [4.2.1](https://github.com/qlik-oss/nebula.js/compare/v4.2.0...v4.2.1) (2023-06-09)
+
+**Note:** Version bump only for package @nebula.js/supernova
+
 # [4.2.0](https://github.com/qlik-oss/nebula.js/compare/v4.1.0...v4.2.0) (2023-05-31)
 
 **Note:** Version bump only for package @nebula.js/supernova